+++ packages/cirrojs/src/index.ts
@@ -5,6 +5,7 @@
 export type { CssOpt } from "./css.ts";
 export { css, genCssFn } from "./css.ts";
 export type { Properties } from "./properties.ts";
-export { getCssRegistry, initCssRegistry } from "./registry.ts";
+export type { Registry } from "./registry.ts";
+export { runWithRegistry } from "./registry.ts";
 export type { AnyRoute, DynamicRoute, Params, ResolvedPage, StaticRoute } from "./router.ts";
 export { expandRoutes, route, urlToFilePath } from "./router.ts";